[pyramid] test suite refactoring
