+++ tests/contrib/pyramid/app/__init__.py
@@ -1,0 +1,1 @@
+from .web import create_app  # noqa
