+++ tests/contrib/pyramid/test_pyramid_autopatch.py
@@ -13,40 +13,24 @@ from wsgiref.simple_server import make_server
 from ...test_tracer import get_dummy_tracer
 from ...util import override_global_tracer
 
-from .test_pyramid import PyramidBase, get_app, custom_exception_view
+#from .test_pyramid import PyramidBase, get_app, custom_exception_view
+from .test_pyramid import PyramidBase
 
 
 class TestPyramidAutopatch(PyramidBase):
-    def setUp(self):
-        self.tracer = get_dummy_tracer()
-        ddtrace.tracer = self.tracer
-
-        config = Configurator()
-        self.rend = config.testing_add_renderer('template.pt')
-        # required to reproduce a regression test
-        config.add_notfound_view(custom_exception_view)
-        app = get_app(config)
-        self.app = webtest.TestApp(app)
+    instrument = False
 
 
 class TestPyramidExplicitTweens(PyramidBase):
-    def setUp(self):
-        self.tracer = get_dummy_tracer()
-        ddtrace.tracer = self.tracer
+    instrument = False
 
-        config = Configurator(settings={'pyramid.tweens': 'pyramid.tweens.excview_tween_factory\n'})
-        self.rend = config.testing_add_renderer('template.pt')
-        # required to reproduce a regression test
-        config.add_notfound_view(custom_exception_view)
-        app = get_app(config)
-        self.app = webtest.TestApp(app)
+    def get_settings(self):
+        return {
+            'pyramid.tweens': 'pyramid.tweens.excview_tween_factory\n',
+        }
 
 
 def _include_me(config):
-    pass
-
-
-def includeme(config):
     pass
 
 
@@ -55,17 +39,3 @@ def test_config_include():
     application is run with ddtrace-run """
     config = Configurator()
     config.include('._include_me')
-
-
-def test_include_conflicts():
-    """ Test that includes do not create conflicts """
-    from ...test_tracer import get_dummy_tracer
-    from ...util import override_global_tracer
-    tracer = get_dummy_tracer()
-    with override_global_tracer(tracer):
-        config = Configurator(settings={'pyramid.includes': 'tests.contrib.pyramid.test_pyramid_autopatch'})
-        app = webtest.TestApp(config.make_wsgi_app())
-        app.get('/', status=404)
-        spans = tracer.writer.pop()
-        assert spans
-        eq_(len(spans), 1)
